move conversions to utils.
